add support for more than 32 TPDOs (actually 64)
the current implementation of CO_get_TPDOmask() and CO_touch_TPDOmask()
use a 32-bit mask thereby only supporting 32 TPDOs.
Switch to a 64-bit unsigned integer so to support up to 64 TPDOs.

+++ CANopen.h
@@ -337,7 +337,7 @@ void CO_touch_TPDOvar(
  * @param pODData Pointer to the data being affected (touched).
  */
 
-uint32_t CO_get_TPDOmask(
+uint64_t CO_get_TPDOmask(
         CO_t                   *CO,
         void                   *pODData);
 
@@ -351,7 +351,7 @@ uint32_t CO_get_TPDOmask(
  */
 void CO_touch_TPDOmask(
         CO_t                   *CO,
-        uint32_t               mask);
+        uint64_t               mask);
 
 #ifdef __cplusplus
 }
